+++ drivers/net/wireless/intel/iwlwifi/mvm/mac80211.c
@@ -2057,7 +2057,7 @@ uintptr_t iwl_mvm_prepare_multicast(struct ieee80211_hw *hw,
 
 	if (pass_all) {
 		cmd->pass_all = 1;
-		return (u64)(unsigned long)cmd;
+		return (uintptr_t)cmd;
 	}
 
 	netdev_hw_addr_list_for_each(addr, mc_list) {
